separate lib from api files and move main MarathonApi to index
