+++ api/artifacts.js
@@ -1,0 +1,6 @@
+module.exports = function createMethods (makeRequest) {
+  return {
+    // /v2/artifacts todo
+    // /v2/artifacts/{path} todo
+  }
+}
