--- lib/artifacts.js
@@ -1,6 +1,0 @@
-module.exports = function createMethods(makeRequest) {
-    return {
-        // /v2/artifacts todo
-        // /v2/artifacts/{path} todo
-    };
-};
