+++ package.json
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/AppliedTrust/marathon-node",
   "dependencies": {
-    "axios": "^1.3.3",
+    "axios": "^1.3.4",
     "eventsource": "^2.0.2"
   },
   "devDependencies": {
